RDKB-7779 : Add logging mechanism to identify atom only reboot
Reason for change: Logging in self heal when atom reboot is triggered
Test Procedure: Verify atom only reboot message in self heal logs
		when atom only reboot is triggered.
Risks: NA

+++ scripts/task_health_monitor.sh
@@ -74,29 +74,33 @@ LIGHTTPD_CONF="/var/lighttpd.conf"
 		resetNeeded psm PsmSsp
 	fi
 
-
-	crTestop=`dmcli eRT getv com.cisco.spvtg.ccsp.CR.Name`
- 	isCRAlive=`echo $crTestop | grep "Can't find destination compo"`
-	if [ "$isCRAlive" != "" ]; then
-		# Retest by querying some other parameter
-		crReTestop=`dmcli eRT getv Device.X_CISCO_COM_DeviceControl.DeviceMode`
- 		isCRAlive=`echo $crReTestop | grep "Can't find destination compo"`
-		  if [ "$isCRAlive" != "" ]; then
-		#	echo "[`getDateTime`] RDKB_PROCESS_CRASHED : CR_process is not running, need to reboot the unit"
-			echo "[`getDateTime`] RDKB_PROCESS_CRASHED : CR_process is not running, need to reboot the unit"
-			vendor=`getVendorName`
-			modelName=`getModelName`
-			CMMac=`getCMMac`
-			timestamp=`getDate`
-			echo "[`getDateTime`] Setting Last reboot reason"
-			reason="CR_crash"
-			rebootCount=1
-			setRebootreason $reason $rebootCount
-			echo "[`getDateTime`] SET succeeded"
-			echo "[`getDateTime`] RDKB_SELFHEAL : <$level>CABLEMODEM[$vendor]:<99000007><$timestamp><$CMMac><$modelName> RM CcspCrSsp process died,need reboot"
-			touch $HAVECRASH
-			rebootNeeded RM "CR"
-		 fi		
+	atomOnlyReboot=`dmesg -n 8 && dmesg | grep -i "Atom only"`
+	if [ x$atomOnlyReboot = x ];then
+		crTestop=`dmcli eRT getv com.cisco.spvtg.ccsp.CR.Name`
+ 		isCRAlive=`echo $crTestop | grep "Can't find destination compo"`
+		if [ "$isCRAlive" != "" ]; then
+			# Retest by querying some other parameter
+			crReTestop=`dmcli eRT getv Device.X_CISCO_COM_DeviceControl.DeviceMode`
+ 			isCRAlive=`echo $crReTestop | grep "Can't find destination compo"`
+		  	if [ "$isCRAlive" != "" ]; then
+		#		echo "[`getDateTime`] RDKB_PROCESS_CRASHED : CR_process is not running, need to reboot the unit"
+				echo "[`getDateTime`] RDKB_PROCESS_CRASHED : CR_process is not running, need to reboot the unit"
+				vendor=`getVendorName`
+				modelName=`getModelName`
+				CMMac=`getCMMac`
+				timestamp=`getDate`
+				echo "[`getDateTime`] Setting Last reboot reason"
+				reason="CR_crash"
+				rebootCount=1
+				setRebootreason $reason $rebootCount
+				echo "[`getDateTime`] SET succeeded"
+				echo "[`getDateTime`] RDKB_SELFHEAL : <$level>CABLEMODEM[$vendor]:<99000007><$timestamp><$CMMac><$modelName> RM CcspCrSsp process died,need reboot"
+				touch $HAVECRASH
+				rebootNeeded RM "CR"
+		 	fi		
+		fi
+	else
+		echo "[`getDateTime`] [RDKB_SELFHEAL] : Atom only reboot is triggered"
 	fi
 
 	PAM_PID=`pidof CcspPandMSsp`
